chore: cleanup acceptance.rs
